requirements.txt
python-telegram-bot==20.7
sqlite3
requests
asyncio